% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_plm.R
\name{brokenstick_prediction}
\alias{brokenstick_prediction}
\title{Title Broken}
\usage{
brokenstick_prediction(
  outcome = "ht",
  time = "time",
  id = "id",
  knots = c(6, 10, 12),
  pred_time = c(2, 4, 6, 8, 10, 12, 14, 16),
  choice = "predicted",
  dataset
)
}
\arguments{
\item{dataset}{}
}
\description{
Title Broken
}
