% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data.R
\docType{data}
\name{epic_train}
\alias{epic_train}
\title{EPIC data for children growth PLM methods}
\format{
\subsection{\code{epic_train}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\subset{train}
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}

\subsection{\code{epic_test}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\subset{train}
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\usage{
epic_train

epic_train
}
\description{
The training subset of data from the EPIC
Report ...

The testing subset of data from the EPIC
Report ...
}
\keyword{datasets}
