% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_simulation.R
\name{gen_all_ranef}
\alias{gen_all_ranef}
\title{Title}
\usage{
gen_all_ranef(
  fullset = margin_mean,
  id = "id",
  knots = c(10, 12, 15),
  degree = 3,
  sigma = 1.04,
  vcov = vcov5,
  seed
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
