% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_simulation.R
\name{gen_ind_ranef}
\alias{gen_ind_ranef}
\title{Title}
\usage{
gen_ind_ranef(
  subset,
  knots = c(10, 12, 15),
  degree = 3,
  sigma = 1.04,
  vcov = vcov5
)
}
\arguments{
\item{vcov}{}
}
\description{
Title
}
