% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_plm.R
\name{linear_brokenstick}
\alias{linear_brokenstick}
\title{Title multivariate linear model}
\usage{
linear_brokenstick(lm_formula = "`.pred` ~ time * sex + baseline", bks_pred)
}
\arguments{
\item{bks_pred}{}
}
\description{
Title multivariate linear model
}
