% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_prediction.R
\name{people_like_me}
\alias{people_like_me}
\title{Title People-Like-Me methods for single testing individual}
\usage{
people_like_me(
  train_data,
  test_data,
  outcome_var = "ht",
  time_var = "time",
  id_var = "id",
  brokenstick_knots = c(5, 12, 15),
  anchor_time = c(2, 5, 10, 11, 12, 12.5, 13, 13.5),
  linear_formula = "ht ~ as.factor(time) * sex + ethnic + genotype + baseline",
  gamlss_formula = "ht ~ cs(time, df = 3)",
  gamlss_sigma = "~ cs(time, df = 1)",
  match_methods = "mahalanobis",
  weight = FALSE,
  match_alpha = NULL,
  match_number = NULL,
  match_plot = FALSE,
  predict_plot = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title People-Like-Me methods for single testing individual
}
