% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_old_plm.R
\name{plm}
\alias{plm}
\title{Title}
\usage{
plm(
  outcome = "ht",
  time = "time",
  id = "id",
  train_data = train,
  test_data = test,
  knots = c(5, 10),
  pred_time = c(4, 6, 8, 12),
  linear_model = "`.pred` ~ time * sex + baseline",
  match_num = NULL,
  match_time = NULL,
  match_methods = "mahalanobis",
  match_alpha = 0.95,
  gamlss_formula = "ht ~ cs(time, df = 3)",
  gamsigma_formula = "~ cs(time, df = 1)",
  match_plot = TRUE,
  predict_plot = TRUE,
  sbj = sbj
)
}
\arguments{
\item{sbj}{}
}
\description{
Title
}
