% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_distance.R, R/99_old_plm.R
\name{plm_ind_plot}
\alias{plm_ind_plot}
\title{Title plot individual matching}
\usage{
plm_ind_plot(quantile, observation, title = NULL)

plm_ind_plot(quantile, observation, title = NULL)
}
\arguments{
\item{title}{}
}
\description{
Title plot individual matching

Title plot individual matching
}
