% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_distance.R
\name{predict_gamlss}
\alias{predict_gamlss}
\title{Title Prediction with GAMLSS model}
\usage{
predict_gamlss(
  matching,
  test_one,
  id_var = "id",
  time_var = "time",
  outcome_var = "ht",
  gamlss_formula = "ht ~ cs(time, df = 3)",
  gamsigma_formula = "~ cs(time, df = 1)",
  weight = FALSE,
  predict_plot = FALSE
)
}
\arguments{
\item{predict_plot}{}
}
\description{
Title Prediction with GAMLSS model
}
