% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_prediction.R
\name{plm_ind_plot}
\alias{plm_ind_plot}
\title{Title plot individual matching}
\usage{
plm_ind_plot(quantile, observation, title = NULL)
}
\arguments{
\item{title}{}
}
\description{
Title plot individual matching
}
